--- docs/code_ref/noise.rst
@@ -1,8 +1,0 @@
-*****
-noise
-*****
-
-This package implements functions for noise level estimation.
-
-.. automodapi:: sunkit_image.noise
-   :no-inheritance-diagram:
