+++ docs/code_ref/utils.rst
@@ -4,4 +4,9 @@ utils
 
 This package contains various utility style functions that can be used throughout ``sunkit-image``.
 
-.. automodapi:: sunkit_image.utils
+All the following are importable using ``sunkit_image.utils``.
+
+.. automodapi:: sunkit_image.utils.noise
+   :no-inheritance-diagram:
+
+.. automodapi:: sunkit_image.utils.utils
