+++ sunkit_image/tests/figure_hashes_py37.json
@@ -2,5 +2,5 @@
     "sunkit_image.tests.test_enhance.test_mgn": "a9ead4186ac2a371175bb27bd8c2cdc8b67772e05993a0241c639c575742bffb",
     "sunkit_image.tests.test_radial.test_fig_nrgf" : "83e0027b78ad32087fadab973977973c69035ab44cfdcf8ab0b1b4d3b69cbe1c",
     "sunkit_image.tests.test_radial.test_fig_fnrgf" : "07663ac57456335ee1d15faf424827fe688a99e3c84da745a862d44af1ef5c3f",
-    "sunkit_image.tests.test_trace.test_occult2_fig": "4f4088b04ca3346b9a9d8ce09bd4254cd9f1bdf15736675356a1c07784fa0082"
+    "sunkit_image.tests.test_trace.test_occult2_fig": "777026dbb32a98b8a36e837f1dc334339e97b8ffec556870447f51d3b77a3e86"
 }
