+++ sunkit_image/utils/noise.py
