+++ sunkit_image/utils/tests/test_noise.py
